handle .length and .items quirks by tagging uses with an __njk signature which is specially handled
custom implementations of the Environment's getitem and getattr are used
to detect these and attempt to take the best action when either of these
are used. we limit this to just lookups with an __njk signature because we
don't want to alter the behaviour of *all* of an app's jinja, and
unfortunately it's not possible to switch Environments as jinja enters into
an .njk file. So we need to specially tag those .length and .items accesses
that originated from an njk file in the preprocess step.

this fixes test_date_input_with_default_items